Stub out Jackson and Gson test classes
Move a few functions out of TestJson and into respective test classes
for JavalinJackson, JavalinGson, and JsonMapper. Other functions may
move as well, but first need to get some feedback on this approach.

Fixed copy/paste/typo bug in `convertLargeStreamToJson`. The
`jsonMapper` passed in is used instead of a hardwired Jackson mapper.
